backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = ""
authors = [
    {name = "Matheus Felipe Pereira Soares",email = "matheusfps27@icloud.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "poetry-plugin-shell (>=1.0.1,<2.0.0)",
    "fastapi (>=0.121.2,<0.122.0)",
    "sqlalchemy (>=2.0.44,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "uvicorn (>=0.38.0,<0.39.0)",
    "requests (>=2.32.5,<3.0.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "logging (>=0.4.9.6,<0.5.0.0)",
    "pydantic[email] (>=2.12.4,<3.0.0)",
    "alembic (>=1.17.1,<2.0.0)",
    "fastapi-pagination (>=0.15.0,<0.16.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)"
]

[tool.poetry.group.dev.dependencies]
pytest = ">=8.3.2,<9.0.0"
httpx = ">=0.27.2,<0.28.0"
pytest-asyncio = ">=0.23.5,<0.24.0"
# Fix de compatibilidade passlib+bcrypt: versões 4.1.x removem __about__ causando erro
# AttributeError: module 'bcrypt' has no attribute '__about__'
# Pin para 4.0.1 até atualização de passlib suportar totalmente 4.1.x
bcrypt = "4.0.1"
requests = "^2.32.5"


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
